requirements.txt
numpy==1.25.2
python-dotenv
openpyxl
pandas
langchain
langchain-community
langchain-openai
langchain-pinecone
openai
pinecone-client
fastapi
uvicorn
sqladmin
alembic
injector
psycopg2-binary

python-jose
passlib[bcrypt]